requirements.txt
asgiref==3.2.7
autopep8==1.5.2
boto3==1.12.48
botocore==1.15.48
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
oauthlib==3.1.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.5.0
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
